+++ packages/eslint-plugin-internal/tests/rules/no-typescript-estree.test.ts
@@ -8,32 +8,80 @@ ruleTester.run('no-typescript-estree-import', rule, {
   invalid: [
     {
       code: "import { foo } from '@typescript-eslint/typescript-estree';",
-      errors: [{ messageId: 'dontImportPackage' }],
+      errors: [
+        {
+          column: 1,
+          endColumn: 60,
+          endLine: 1,
+          line: 1,
+          messageId: 'dontImportPackage',
+        },
+      ],
       output: "import { foo } from '@typescript-eslint/utils';",
     },
     {
       code: "import foo from '@typescript-eslint/typescript-estree';",
-      errors: [{ messageId: 'dontImportPackage' }],
+      errors: [
+        {
+          column: 1,
+          endColumn: 56,
+          endLine: 1,
+          line: 1,
+          messageId: 'dontImportPackage',
+        },
+      ],
       output: "import foo from '@typescript-eslint/utils';",
     },
     {
       code: "import * as foo from '@typescript-eslint/typescript-estree';",
-      errors: [{ messageId: 'dontImportPackage' }],
+      errors: [
+        {
+          column: 1,
+          endColumn: 61,
+          endLine: 1,
+          line: 1,
+          messageId: 'dontImportPackage',
+        },
+      ],
       output: "import * as foo from '@typescript-eslint/utils';",
     },
     {
       code: "import { foo } from '@typescript-eslint/types';",
-      errors: [{ messageId: 'dontImportPackage' }],
+      errors: [
+        {
+          column: 1,
+          endColumn: 48,
+          endLine: 1,
+          line: 1,
+          messageId: 'dontImportPackage',
+        },
+      ],
       output: "import { foo } from '@typescript-eslint/utils';",
     },
     {
       code: "import foo from '@typescript-eslint/types';",
-      errors: [{ messageId: 'dontImportPackage' }],
+      errors: [
+        {
+          column: 1,
+          endColumn: 44,
+          endLine: 1,
+          line: 1,
+          messageId: 'dontImportPackage',
+        },
+      ],
       output: "import foo from '@typescript-eslint/utils';",
     },
     {
       code: "import * as foo from '@typescript-eslint/types';",
-      errors: [{ messageId: 'dontImportPackage' }],
+      errors: [
+        {
+          column: 1,
+          endColumn: 49,
+          endLine: 1,
+          line: 1,
+          messageId: 'dontImportPackage',
+        },
+      ],
       output: "import * as foo from '@typescript-eslint/utils';",
     },
   ],