+++ packages/eslint-plugin/tests/rules/prefer-nullish-coalescing.test.ts
@@ -906,35 +906,35 @@ do {} while ((x ||= 'foo'));
     `,
     `
 declare let x: string | null | undefined;
-for (; x || 'foo'; ) {}
+for (; x || 'foo';) {}
     `,
     `
 declare let x: string | null | undefined;
-for (; (x ||= 'foo'); ) {}
+for (; (x ||= 'foo');) {}
     `,
     `
 declare let x: number | null | undefined;
-for (; x || 'foo'; ) {}
+for (; x || 'foo';) {}
     `,
     `
 declare let x: number | null | undefined;
-for (; (x ||= 'foo'); ) {}
+for (; (x ||= 'foo');) {}
     `,
     `
 declare let x: boolean | null | undefined;
-for (; x || 'foo'; ) {}
+for (; x || 'foo';) {}
     `,
     `
 declare let x: boolean | null | undefined;
-for (; (x ||= 'foo'); ) {}
+for (; (x ||= 'foo');) {}
     `,
     `
 declare let x: object | null | undefined;
-for (; x || 'foo'; ) {}
+for (; x || 'foo';) {}
     `,
     `
 declare let x: object | null | undefined;
-for (; (x ||= 'foo'); ) {}
+for (; (x ||= 'foo');) {}
     `,
     `
 declare let x: string | null | undefined;
@@ -6591,7 +6591,7 @@ do {} while ((x ??= 'foo'));
     {
       code: `
 declare let x: string | null | undefined;
-for (; x || 'foo'; ) {}
+for (; x || 'foo';) {}
       `,
       errors: [
         {
@@ -6605,7 +6605,7 @@ for (; x || 'foo'; ) {}
               messageId: 'suggestNullish',
               output: `
 declare let x: string | null | undefined;
-for (; x ?? 'foo'; ) {}
+for (; x ?? 'foo';) {}
       `,
             },
           ],
@@ -6617,7 +6617,7 @@ for (; x ?? 'foo'; ) {}
     {
       code: `
 declare let x: string | null | undefined;
-for (; (x ||= 'foo'); ) {}
+for (; (x ||= 'foo');) {}
       `,
       errors: [
         {
@@ -6631,7 +6631,7 @@ for (; (x ||= 'foo'); ) {}
               messageId: 'suggestNullish',
               output: `
 declare let x: string | null | undefined;
-for (; (x ??= 'foo'); ) {}
+for (; (x ??= 'foo');) {}
       `,
             },
           ],
@@ -6643,7 +6643,7 @@ for (; (x ??= 'foo'); ) {}
     {
       code: `
 declare let x: number | null | undefined;
-for (; x || 'foo'; ) {}
+for (; x || 'foo';) {}
       `,
       errors: [
         {
@@ -6657,7 +6657,7 @@ for (; x || 'foo'; ) {}
               messageId: 'suggestNullish',
               output: `
 declare let x: number | null | undefined;
-for (; x ?? 'foo'; ) {}
+for (; x ?? 'foo';) {}
       `,
             },
           ],
@@ -6669,7 +6669,7 @@ for (; x ?? 'foo'; ) {}
     {
       code: `
 declare let x: number | null | undefined;
-for (; (x ||= 'foo'); ) {}
+for (; (x ||= 'foo');) {}
       `,
       errors: [
         {
@@ -6683,7 +6683,7 @@ for (; (x ||= 'foo'); ) {}
               messageId: 'suggestNullish',
               output: `
 declare let x: number | null | undefined;
-for (; (x ??= 'foo'); ) {}
+for (; (x ??= 'foo');) {}
       `,
             },
           ],
@@ -6695,7 +6695,7 @@ for (; (x ??= 'foo'); ) {}
     {
       code: `
 declare let x: boolean | null | undefined;
-for (; x || 'foo'; ) {}
+for (; x || 'foo';) {}
       `,
       errors: [
         {
@@ -6709,7 +6709,7 @@ for (; x || 'foo'; ) {}
               messageId: 'suggestNullish',
               output: `
 declare let x: boolean | null | undefined;
-for (; x ?? 'foo'; ) {}
+for (; x ?? 'foo';) {}
       `,
             },
           ],
@@ -6721,7 +6721,7 @@ for (; x ?? 'foo'; ) {}
     {
       code: `
 declare let x: boolean | null | undefined;
-for (; (x ||= 'foo'); ) {}
+for (; (x ||= 'foo');) {}
       `,
       errors: [
         {
@@ -6735,7 +6735,7 @@ for (; (x ||= 'foo'); ) {}
               messageId: 'suggestNullish',
               output: `
 declare let x: boolean | null | undefined;
-for (; (x ??= 'foo'); ) {}
+for (; (x ??= 'foo');) {}
       `,
             },
           ],
@@ -6747,7 +6747,7 @@ for (; (x ??= 'foo'); ) {}
     {
       code: `
 declare let x: object | null | undefined;
-for (; x || 'foo'; ) {}
+for (; x || 'foo';) {}
       `,
       errors: [
         {
@@ -6761,7 +6761,7 @@ for (; x || 'foo'; ) {}
               messageId: 'suggestNullish',
               output: `
 declare let x: object | null | undefined;
-for (; x ?? 'foo'; ) {}
+for (; x ?? 'foo';) {}
       `,
             },
           ],
@@ -6773,7 +6773,7 @@ for (; x ?? 'foo'; ) {}
     {
       code: `
 declare let x: object | null | undefined;
-for (; (x ||= 'foo'); ) {}
+for (; (x ||= 'foo');) {}
       `,
       errors: [
         {
@@ -6787,7 +6787,7 @@ for (; (x ||= 'foo'); ) {}
               messageId: 'suggestNullish',
               output: `
 declare let x: object | null | undefined;
-for (; (x ??= 'foo'); ) {}
+for (; (x ??= 'foo');) {}
       `,
             },
           ],